requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.3
Django==4.2.5
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-environ==0.11.2
django-grappelli==3.0.8
django-js-asset==2.1.0
django-mail-queue==3.2.5
django-richtextfield==1.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.0
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
